Advanced Smart Contracts Lab 2

+++ contracts/SomeLibrary.sol
@@ -1,13 +1,13 @@
 pragma solidity ^0.5.0;
 
 contract SomeLibrary {
-    uint public version = 1;
+    uint256 public version = 1;
 
-    function setVersion(uint newVersion) public {
+    function setVersion(uint256 newVersion) public {
         version = newVersion;
     }
 
-    function getMsgSender() public view returns(address){
+    function getMsgSender() public view returns (address) {
         return msg.sender;
     }
 }
